Make the EntityIdentifier Constructor Easier to Use
And also require callers to have a Source implementation present. This
conforms to the principle of "make it hard to do the wrong thing; easy to
do the right thing".

Pushing with warning because the "unused code" warning is the result of the
code not being used outside of test code. It will get used in subsequent
commits.